requirements.txt
# Core dependencies
python-dotenv==1.0.1
pandas==2.2.0
httpx==0.26.0
beautifulsoup4==4.12.3
playwright==1.41.2
pymupdf==1.24.0
lxml==5.1.0

# Utilities
unicodedata2==15.1.0
python-dateutil==2.8.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==24.2.0
flake8==7.0.0
mypy==1.8.0